+++ lib/pairings_engine_web/router.ex
@@ -49,6 +49,7 @@ defmodule PairingsEngineWeb.Router do
 
     get "/t/:id/print/players", PrintController, :player_list
     get "/t/:id/print/cards", PrintController, :player_cards
+    get "/t/:id/print/card/:player_id", PrintController, :player_card
     get "/t/:id/print/placecards", PrintController, :place_cards
     get "/t/:id/print/pairings", PrintController, :pairing_list
     get "/t/:id/print/pairings-alpha", PrintController, :pairing_alpha